+++ Source/Core/Transforms.js
@@ -564,6 +564,10 @@ define([
         }
         //>>includeEnd('debug');
 
+        if (!defined(result)) {
+            result = new Cartesian2();
+        }
+
         var tmp = pointToWindowCoordinatesTemp;
 
         Matrix4.multiplyByVector(modelViewProjectionMatrix, Cartesian4.fromElements(point.x, point.y, point.z, 1, tmp), tmp);